refactor: :recycle: improve IMU_HZ Kconfig
Additionally, add a basic state machine frame.

+++ include/lib/imu.h
@@ -22,7 +22,7 @@
 #if !defined(CONFIG_LSM6DSO_TRIGGER)
 /**
  * @brief Poll the IMU for new data.
- * 
+ *
  * Function fetches new data from the IMU device.
  *
  * @param dev Pointer to the IMU device
@@ -34,16 +34,16 @@ int imu_poll(const struct device *dev);
 
 /**
  * @brief Set the IMU sampling frequency.
- * 
+ *
  * Function sets the sampling frequency of the IMU device.
- * 
+ *
  * @param dev Pointer to the IMU device
- * 
+ *
  * @param sampling_rate_hz The desired sampling rate in Hertz.
- * 
+ *
  * @retval 0 on success, negative error code on failure.
  */
-int imu_set_sampling_freq(const struct device *dev, float sampling_rate_hz);
+int imu_set_sampling_freq(const struct device *dev, int sampling_rate_hz);
 
 /**
 * @brief Initialize the IMU for the AURORA application.
